requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cloudinary==1.30.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
